implementation of profile usecase and combine profile editor usecase
