+++ FunchApp/Domain/UseCase/ProfileScene/ProfileUseCase.swift
@@ -1,5 +1,5 @@
 //
-//  ProfileSceneUseCase.swift
+//  File.swift
 //  FunchApp
 //
 //  Created by Geon Woo lee on 1/20/24.
@@ -7,6 +7,22 @@
 
 import Foundation
 
-final class ProfileSceneUseCase {
+final class ProfileUseCase {
+    private let profileRepository: ProfileRepository
     
+    init() {
+        self.profileRepository = ProfileRepository()
+    }
+    
+    /// 본인 프로필 생성
+    func createProfile(createUserQuery: CreateUserQuery, completion: @escaping (Result<Void, Error>) -> Void) {
+        profileRepository.createProfile(createUserQuery: createUserQuery) { result in
+            switch result {
+            case .success(let success):
+                completion(.success(()))
+            case .failure(let failure):
+                completion(.success(()))
+            }
+        }
+    }
 }
